Separate non-Windows mount code from common code
